Adding GitFileCounter class for counting file changes related to bugs.
This commit adds a new class GitFileCounter to the project. The class includes methods for reading a list of keys from a text file, finding file changes related to these keys in a Git commit log file, and writing the results to a CSV file. The output file includes the names of files and the number of post-release bugs associated with each file.
This can be customised for listing down information of both post and pre release bugs.