Pushing collected metrics until now
1. Post-release defects:
Write a custom Java program that reads in a list of keys associated with post-release bugs from a file, counts the number of changes to each file associated with the given keys in a commit log file, and writes the output to a CSV file.

2. Traditional Process Metrics:
a. Collected total number of changes, total code churn using git log command with the options --numstat and --pretty.
b. Collected pre-release defects by following the same approach as that of post-release bugs.

3. Exception flow metrics:
These were collected using https://github.com/guipadua/JTratch command-line tool.

4. Traditional product metrics:
To extract traditional product metrics for Hibernate-orm 5.2.0 we used Understand tool, from SciTools website. Created a new project and configure the project directory and then clicked on the analyze option. Once the analysis is complete, we could view the results in the Metrics tab of Understand. The results will show the values of selected metrics for each file, function, class, and other components of the project, which can be used to understand and improve the code quality. Later, this was exported to a CSV file for further analysis.